refactor/6-move-custom_scripts-to-weekly_working_hours.py
