--- hr_addon/custom_scripts/__init__.py
@@ -1,3 +1,0 @@
-
-__version__ = '0.0.1'
-
