--- hr_addon/custom_scripts/custom_python/__init__.py
@@ -1,3 +1,0 @@
-
-__version__ = '0.0.1'
-
