+++ hr_addon/hooks.py
@@ -206,7 +206,7 @@ scheduler_events = {
 		"hr_addon.hr_addon.doctype.workday.workday.generate_workdays_scheduled_job"
 	],
     "yearly": [
-        "hr_addon.custom_scripts.custom_python.weekly_working_hours.set_from_to_dates",
+        "hr_addon.hr_addon.doctype.weekly_working_hours.weekly_working_hours.set_from_to_dates",
 	],
 	"daily": [
 		"hr_addon.hr_addon.doctype.hr_addon_settings.hr_addon_settings.send_work_anniversary_notification"
