+++ hr_addon/public/js/list_view.js
@@ -4,7 +4,7 @@ frappe.listview_settings["Weekly Working Hours"] = {
 		
 			list_view.page.add_button(__("Update Year"), function () {
 				frappe.call({
-					method: "hr_addon.custom_scripts.custom_python.weekly_working_hours.set_from_to_dates",
+					method: "hr_addon.hr_addon.doctype.weekly_working_hours.weekly_working_hours.set_from_to_dates",
 					args: {},
 					callback(r) {}
 				});
